Initial implemention of procedural quote! core
This doesn't move most or all of the functionality into the internal
macro, but rather tries to move the pieces which are easier to.

This current implementation is unfortunately not quite where it should
be to be merged, although it appears to have slightly better runtime
performance than the existing quote! macro.